+++ PerseusPluginLib/Mods/ExpandSiteTable.cs
@@ -50,8 +50,8 @@ namespace PerseusPluginLib.Mods {
 			int[] validNumCols = GetValidNumCols(mdata.NumericColumnNames, allPrefixes);
 			int nrows = mdata.RowCount * allSuffixes.Length;
 			int ncols = normalIndices.Length + allPrefixes.Length;
-			float[,] data = new float[nrows, ncols];
-			float[,] quality = new float[nrows, ncols];
+			double[,] data = new double[nrows, ncols];
+			double[,] quality = new double[nrows, ncols];
 			bool[,] imputed = new bool[nrows, ncols];
 			List<double[]> numCols = new List<double[]>();
 			for (int i = 0; i < validNumCols.Length; i++) {
@@ -85,13 +85,13 @@ namespace PerseusPluginLib.Mods {
 					count++;
 					int rowInd = i * mdata.RowCount + j;
 					for (int k = 0; k < normalIndices.Length; k++) {
-						data[rowInd, k] = (float)mdata.Values.Get(j, normalIndices[k]);
-						quality[rowInd, k] = (float)mdata.Quality.Get(j, normalIndices[k]);
+						data[rowInd, k] = mdata.Values.Get(j, normalIndices[k]);
+						quality[rowInd, k] = mdata.Quality.Get(j, normalIndices[k]);
 						imputed[rowInd, k] = mdata.IsImputed[j, normalIndices[k]];
 					}
 					for (int k = 0; k < allPrefixes.Length; k++) {
-						data[rowInd, normalIndices.Length + k] = (float)mdata.Values.Get(j, colInds[k, i]);
-						quality[rowInd, normalIndices.Length + k] = (float)mdata.Quality.Get(j, colInds[k, i]);
+						data[rowInd, normalIndices.Length + k] = mdata.Values.Get(j, colInds[k, i]);
+						quality[rowInd, normalIndices.Length + k] = mdata.Quality.Get(j, colInds[k, i]);
 						imputed[rowInd, normalIndices.Length + k] = mdata.IsImputed[j, colInds[k, i]];
 					}
 					for (int k = 0; k < validNumCols.Length; k++) {
